Change PlayerSummary.GameID to uint64

+++ profile.go
@@ -48,7 +48,7 @@ type PlayerSummary struct {
 	LocCountryCode    string  `json:"loccountrycode"`
 	LocStateCode      string  `json:"locstatecode"`
 	LocCityID         uint32  `json:"loccityid"`
-	GameID            uint32  `json:"gameid,string"`
+	GameID            uint64  `json:"gameid,string"`
 	GameServerIP      string  `json:"gameserverip"`
 	GameExtraInfo     string  `json:"gameextrainfo"`
 }
